move the code from nanoagent to this pubic repository this faciliteates the use of the package in nanobox as well as nanoagent.
A few minor changes had to be added to remove the hard dependency
on viper as well as lumber.
TODO
nanoagent needs to be update to depend on this external repo instead of
its internal one. and the Initialize() has to be called in nanoagent
before a api call is made to the docker-cleint